code review

+++ ContactPicker/THContactPickerViewController.m
@@ -151,7 +151,7 @@
 - (void)contactPickerDidRemoveContact:(id)contact {
     [self.selectedContacts removeObject:contact];
 
-    int index = [self.contacts indexOfObject:contact];
+    NSInteger index = [self.contacts indexOfObject:contact];
     UITableViewCell *cell = [self.tableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:index inSection:0]];
     cell.accessoryType = UITableViewCellAccessoryNone;
 }
